Keep method name symbols for alive methods alive
Liquids has a security test which asserts that there is no growth in
entry count for the symbol table when parsing and rendering. It may
happen that some method name symbols used in the call graph of said
security test is GC'ed prior to the test running, but are recreated
during the execution of the test. This causes test flakiness as
different test orders affect what method name symbols are alive when the
security test runs.

This fix keeps a reference to the method name symbol inside
InternalMethod. This keeps the symbol alive and present in the symbol
table.

Some symbols are GC roots in MRI and I don't think there is something
equivalent in Truffle. Let's see if this commit alone is enough to
address Liquid's issue.