Make macros around TypedData_Wrap_Struct work better
> struct RData *RDATA(VALUE value);
> #define RTYPEDDATA(value) ((struct RTypedData *)RDATA(value))

This looks bad. RDATA returns a Ruby object that handles Truffle
interop messages. The only READ message it works on is 'data',
which it implements using the data holder hidden ivar.

Accessing other fields in struct RData sends different message
and was causing weird failures, especially due to the cast to
RTypedData. Fields in RTypedData were mapped to fields in
struct RData because the field names are compiled down into
offsets, I guess.

# Conflicts:
#	lib/truffle/truffle/cext.rb